Began training proof of concept
Began training and tuning on hyper params. Not seeing myself get above
val accuacy of 27.5%. With 40 classes, that could be good. But what are
the baselines?

If the model predicted randomly... Accuracy would be 1/40... If the
model just predicts Michael every time... the accuracy wouldnt be too
far from 27%.

Time to simplify # classes & randomly oversample.